Init toolbox development container
Initializes a development toolbox container intended to provide all
necessary tooling for the repository in a declarative way in an Ubuntu
environment. This is a fairly stock version I like to keep handy as a
baseline for new projects.

Note: because distrobox takes advantage of your $HOME directory and its
contents, and because the vscode remote containers extension doesn't
play perfectly, creating and entering the toolbox will create a
directory and file in your $HOME to make attaching to the container
seamless within vscode.