Init super-linter via xc
Initializes the use of superlinter (run locally) via Tasks entry in the
top-level README.

n.b. even the slim super-linter container is a >5GB container image.
Since this is a private repository in a personal account, I'm going to
avoid creating CI actions up front. We can hook this into the
development workflow with a git hook for the time being.