Add direnv to dev toolbox
Adds direnv to the dev toolbox. This enables leveraging automatic
loading of environment variables. I don't have a great story for how to
set this up if the host doesn't also already use it, as we effectively
need a new layer to hook directly into the shell. Definitely something
to poke at in the future.